**10 torrents**
Before:
1) This comparison https://github.com/qbittorrent/qBittorrent/blob/master/src/webui/www/public/scripts/dynamicTable.js#L465 is done 247 times * 10 torrents = 2470 times / second.
2) Updates all rows and columns in the transfer list.
3) Time of execution (average) of this call https://github.com/qbittorrent/qBittorrent/blob/master/src/webui/www/public/scripts/client.js#L177 **30 ms.** (it's called each second so it's not negligible)
After:
1) 11 times * 10 torrents = 110 times / second.
2) Only updates changed values. Fixed in this line https://github.com/qbittorrent/qBittorrent/pull/3467/files#diff-1d55956245e67ad3c0b3eb4c48509ecdR465
3) **5 ms.**

**100 torrents**
Before:
1) 247 times * 100 torrents = 24700 times / second.
2) Updates all rows and columns in the transfer list.
3) **85 ms.**
After:
1) 11 times * 100 torrents = 1100 times / second.
2) Only updates changed values.
3) **20 ms.**